server core 2016 complete

+++ Windows/scripts/Shutdown.ps1
@@ -1,6 +1,6 @@
 param
 (
-    [parameter()]
+    [Parameter()]
     [switch] $UseCoreOSWorkaround = $false
 )
 
@@ -8,18 +8,34 @@ if($UseCoreOSWorkaround)
 {
     Write-Warning "Cleaning up PowerShell profile workaround for startup items"
 
-    Remove-ItemProperty `
-            -Path "HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Winlogon" `
-            -Name Shell
-    Remove-Item -Force $PROFILE
+    if(Test-Path 'HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Winlogon')
+    {
+        Remove-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Winlogon" -Name Shell
+    }
+
+    if(Test-Path $PROFILE)
+    {
+        Remove-Item -Force $PROFILE
+    }
 }
 
-Remove-Item -Force -recurse -path $Env:AppData\BoxstarterSection
 
-Set-Item -Path WSMan:\localhost\Service\AllowUnencrypted -Value false
-Set-Item -Path WSMan:\localhost\Service\Auth\Basic       -Value false
-netsh advfirewall firewall set rule name="WinRM-HTTP" new action=block
+if(Test-Path $Env:AppData\BoxstarterSection)
+{
+    Remove-Item -Force -recurse -path $Env:AppData\BoxstarterSection
+}
 
-Start-Process -FilePath "C:\windows\system32\sysprep\sysprep.exe" -ArgumentList "/generalize /oobe /unattend:C:/Windows/Panther/Unattend/unattend.xml /quiet /shutdown"
-
-   
+try
+{
+    Set-Item -Path WSMan:\localhost\Service\AllowUnencrypted -Value false
+    Set-Item -Path WSMan:\localhost\Service\Auth\Basic       -Value false
+    netsh advfirewall firewall set rule name="WinRM-HTTP" new action=block
+}
+catch
+{
+#Do nothing
+}
+finally
+{
+    Start-Process -FilePath "C:\windows\system32\sysprep\sysprep.exe" -ArgumentList "/generalize /oobe /unattend:C:\Windows\Panther\Unattend\unattend.xml /shutdown" -Wait
+}
